refactoring tests block
